init: define our own service to load touch panel and display modules, fixing the race condition
Actually, if we just place these modules inside /vendor/lib/modules, TWRP won't load them sequentially by reading modules.load. This causes the touch modules to load first, before the msm_drm module. That breaks the binding between the touch panel and the actual display panel - the famous 'Chipone and Himax touch issue' - which was considered unfixable for a year, but was finally fixed by a random guy on GitHub (yeah, that’s me: @ravindu644)

feat: added an init service to log the dmesg to a /tmp for debugging purposes

Signed-off-by: ravindu644 <droidcasts@protonmail.com>