CRITICAL: kernel: cmdline: fix firmware class path
This was the second issue behind the famous “Chipone and Himax” touchscreen bug.

While analyzing dmesg, I found that the Himax touch driver was struggling to load its firmware.

When I checked `/sys/module/firmware_class/parameters/path`, I discovered that it was set only to `/vendor/firmware_mnt/image`.

However, the actual touch firmware is located in `/vendor/firmware/firmware`.

Because of this, the kernel never knew where to find the touch firmware, resulting in a broken touchscreen.